requirements.txt
langchain
openai
chromadb
faiss-cpu
tiktoken
pypdf
python-dotenv
